+++ src/js/load-package.ts
@@ -278,10 +278,16 @@ async function downloadPackage(
   let installBaseUrl: string;
   if (IN_NODE) {
     installBaseUrl = API.config.packageCacheDir;
-    // ensure that the directory exists before trying to download files into it
-    await nodeFsPromisesMod.mkdir(API.config.packageCacheDir, {
-      recursive: true,
-    });
+    // Ensure that the directory exists before trying to download files into it.
+    try {
+      // Check if the `installBaseUrl` directory exists
+      await nodeFsPromisesMod.stat(installBaseUrl); // Use `.stat()` which works even on ASAR archives of Electron apps, while `.access` doesn't.
+    } catch {
+      // If it doesn't exist, make it. Call mkdir() here only when necessary after checking the existence to avoid an error on read-only file systems. See https://github.com/pyodide/pyodide/issues/4736
+      await nodeFsPromisesMod.mkdir(installBaseUrl, {
+        recursive: true,
+      });
+    }
   } else {
     installBaseUrl = API.config.indexURL;
   }