+++ docs/project/changelog.md
@@ -16,6 +16,12 @@ myst:
 
 ## Unreleased
 
+- {{ Enhancement }} `pyodide.loadPackage` now checks if the cache directory exists and calls `mkdir` only when it doesn't to avoid an error on read-only file systems in Node.js environment.
+  {pr}`4738`
+
+- {{ Fix }} pyodide-build now use response file when passing list of exported symbols to `emcc`.
+  This Fixes "Argument list too long" error.
+
 - {{ Fix }} Pass through `-E` (command mode) arguments in CMake wrapper {pr}`4705`.
 
 - {{ Fix }} Fix exception handling in dynamic linking of int64 functions {pr}`4698`.